Update dependencies
Update dependencies to latest versions, except `middleman` which is not
working correctly.

Update minimum Ruby version to 2.4.0.
Remove older Ruby version (2.3.3) from Travis CI configuration because
of incompatibilities with the newer versions of the dependencies (a
newer Ruby version is already being used to generate the site).

The generated content has not changed.

Signed-off-by: thc202 <thc202@gmail.com>